# api/src/main/resources/META-INF/swagger/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Bisq HTTP API",
        "version": "0.0.1"
    },
    "security": [
        {
            "authorization": []
        }
    ],
...
    "paths": {
        "/api/v1/payment-accounts/{id}": {
            "delete": {
                "tags": [
                    "payment-accounts"
                ],
                "summary": "Remove payment account",
                "description": "This is EXPERIMENTAL FEATURE! Run it at your own risk! Requires --enableHttpApiExperimentalFeatures flag at startup.",
                "operationId": "removeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "default response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "authorization": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}